requirements.txt
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.4
django-axes==8.3.1
django-csp==4.0
django-ratelimit==4.1.0
docstring_parser==0.18.0
et_xmlfile==2.0.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
openpyxl==3.1.5
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.11
pydantic==2.13.1
pydantic_core==2.46.1
PyPDF2==3.0.1
python-decouple==3.8
python-magic-bin==0.4.14
qrcode==8.2
reportlab==4.4.10
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
